# swagger.json
{
  "consumes": [
    "application/json",
    "application/xml"
  ],
  "produces": [
    "application/json",
    "application/xml"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "BrightBird API.",
    "contact": {
      "email": "venus@ipfsconmmunity.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "0.0.1"
  },
  "host": "192.168.200.103:12356",
  "basePath": "/api/v1",
  "paths": {
...
    "/logs": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "log"
        ],
        "summary": "get all logs in pod.",
        "operationId": "podLogReq",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TestID",
            "description": "testid of task",
            "name": "testID",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "PodName",
            "description": "pod name",
            "name": "podLog",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "logResp",
            "schema": {
              "$ref": "#/definitions/logResp"
            }
          },
          "503": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
    "/logs/pods/{testid}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "log"
        ],
        "summary": "List all pod names in test.",
        "operationId": "listPodsInTest",
        "parameters": [
          {
            "type": "string",
            "description": "test id",
            "name": "testid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "stringArr",
            "schema": {
              "$ref": "#/definitions/stringArr"
            }
          },
          "503": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
...
    "/task": {
      "get": {
        "description": "Get task by condition",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "task"
        ],
        "operationId": "getTaskReq",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TestID",
            "name": "testID",
            "in": "query"
          },
          {
            "type": "string",
            "name": "ID",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "task",
            "schema": {
              "$ref": "#/definitions/task"
            }
          },
          "503": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "StepState": {
      "type": "string",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
...
    "TestId": {
      "type": "string",
      "x-go-package": "github.com/ipfs-force-community/brightbird/types"
    },
...
    "apiError": {
      "description": "APIError",
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-name": "APIError",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
...
    "logResp": {
      "description": "LogResp",
      "type": "object",
      "properties": {
        "logs": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Logs"
        },
        "podName": {
          "type": "string",
          "x-go-name": "PodName"
        },
        "steps": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/stepLog"
          },
          "x-go-name": "Steps"
        }
      },
      "x-go-name": "LogResp",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
...
    "stepLog": {
      "description": "StepLog",
      "type": "object",
      "properties": {
        "instanceName": {
          "type": "string",
          "x-go-name": "InstanceName"
        },
        "logs": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Logs"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "state": {
          "$ref": "#/definitions/StepState"
        }
      },
      "x-go-name": "StepLog",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
    "stringArr": {
      "description": "StringArray",
      "type": "array",
      "items": {
        "type": "string"
      },
      "x-go-name": "StringArray",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
    "task": {
      "description": "Task",
      "type": "object",
      "properties": {
        "createTime": {
          "description": "创建时间",
          "type": "string",
          "format": "int64",
          "x-go-name": "CreateTime"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "inheritVersions": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "InheritVersions"
        },
        "jobId": {
          "type": "string",
          "x-go-name": "JobId"
        },
        "logs": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Logs"
        },
        "modifiedTime": {
          "description": "最后修改时间",
          "type": "string",
          "format": "int64",
          "x-go-name": "ModifiedTime"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "podName": {
          "type": "string",
          "x-go-name": "PodName"
        },
        "retryTime": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RetryTime"
        },
        "state": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "State"
        },
        "testFlowId": {
          "type": "string",
          "x-go-name": "TestFlowId"
        },
        "testId": {
          "$ref": "#/definitions/TestId"
        },
        "versions": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "CommitMap"
        }
      },
      "x-go-name": "Task",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
...
  },
...
}